+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version history
 
 ## 0.1.0
+- Verify required `srb`, `bundle` during extension startup.
 - Initial version, forked from official 0.3.41.